@@ -1,6 +1,6 @@
 import { Component } from '@angular/core';
 import { Store } from '@ngrx/store';
-import { Observable } from 'rxjs/Observable';
+import { Observable } from 'rxjs';
 import { AutoUpdateSettings } from 'uhk-common';
 
 import { AppState, getAutoUpdateSettings, getCheckingForUpdate } from '../../../store';
