@@ -1,10 +1,9 @@
-import { AfterViewInit, Component, OnInit, OnDestroy, ViewChildren, QueryList } from '@angular/core';
+import { Component, OnInit, OnDestroy } from '@angular/core';
 import { Store } from '@ngrx/store';
 import { AppState, getUserConfiguration } from '../../../store';
 import { SetUserConfigurationValueAction } from '../../../store/actions/user-config';
 import { SliderPips } from '../../slider-wrapper/slider-wrapper.component';
-import { Observable } from 'rxjs/Observable';
-import { Subscription } from 'rxjs/Subscription';
+import { Observable, Subscription } from 'rxjs';
 import { UserConfiguration } from 'uhk-common';
 
 @Component({
@@ -26,7 +25,7 @@ export class LEDBrightnessComponent implements OnInit, OnDestroy {
         stepped: true
     };
 
-    private userConfig$: Store<UserConfiguration>;
+    private userConfig$: Observable<UserConfiguration>;
     private userConfigSubscription: Subscription;
 
     constructor(private store: Store<AppState>) {}
