@@ -14,8 +14,7 @@ import {
 import { animate, keyframes, state, style, transition, trigger } from '@angular/animations';
 
 import { Store } from '@ngrx/store';
-import { Observable } from 'rxjs/Observable';
-import { BehaviorSubject } from 'rxjs/BehaviorSubject';
+import { BehaviorSubject, Observable } from 'rxjs';
 import { combineLatest, map } from 'rxjs/operators';
 
 import {
