@@ -13,8 +13,7 @@ import {
     ViewChild
 } from '@angular/core';
 
-import { Observable } from 'rxjs/Observable';
-import { of } from 'rxjs/observable/of';
+import { Observable, of } from 'rxjs';
 import { map } from 'rxjs/operators';
 import { Store } from '@ngrx/store';
 
@@ -109,7 +108,7 @@ export class SvgKeyboardWrapComponent implements OnInit, OnChanges {
         this.tooltipData = {
             posTop: 0,
             posLeft: 0,
-            content: Observable.of([]),
+            content: of([]),
             show: false
         };
     }
@@ -302,7 +301,7 @@ export class SvgKeyboardWrapComponent implements OnInit, OnChanges {
                     value: SecondaryRoleAction[keystrokeAction.secondaryRoleAction]
                 });
             }
-            return Observable.of(content);
+            return of(content);
         } else if (keyAction instanceof MouseAction) {
             const mouseAction: MouseAction = keyAction;
             const content: NameValuePair[] =
@@ -316,7 +315,7 @@ export class SvgKeyboardWrapComponent implements OnInit, OnChanges {
                         value: camelCaseToSentence(MouseActionParam[mouseAction.mouseAction])
                     }
                 ];
-            return Observable.of(content);
+            return of(content);
         } else if (keyAction instanceof PlayMacroAction) {
             const playMacroAction: PlayMacroAction = keyAction;
             return this.store
