@@ -1,7 +1,7 @@
 import { Injectable } from '@angular/core';
 import { Router } from '@angular/router';
 
-import { Actions, Effect } from '@ngrx/effects';
+import { Actions, Effect, ofType } from '@ngrx/effects';
 import { Store, Action } from '@ngrx/store';
 import { map, pairwise, tap, withLatestFrom } from 'rxjs/operators';
 
@@ -15,8 +15,8 @@ import { findNewItem } from '../../util';
 export class MacroEffects {
 
     @Effect({ dispatch: false }) remove$: any = this.actions$
-        .ofType<Macros.RemoveMacroAction>(Macros.ActionTypes.Remove)
         .pipe(
+            ofType<Macros.RemoveMacroAction>(Macros.ActionTypes.Remove),
             tap(action => this.store.dispatch(new Keymaps.CheckMacroAction(action.payload))),
             withLatestFrom(this.store.select(getMacros)),
             map(([action, macros]) => macros),
@@ -31,14 +31,15 @@ export class MacroEffects {
         );
 
     @Effect({ dispatch: false }) addOrDuplicate$: any = this.actions$
-        .ofType(Macros.ActionTypes.Add, Macros.ActionTypes.Duplicate)
         .pipe(
+            ofType<Macros.AddMacroAction | Macros.DuplicateMacroAction>(
+                Macros.ActionTypes.Add, Macros.ActionTypes.Duplicate),
             withLatestFrom(this.store.select(getMacros)
                 .pipe(
                     pairwise()
                 )
             ),
-            map(([action, latest]) => ([action, latest[0], latest[1]])),
+            map(([action, latest]: [Action, Macro[][]]) => ([action, latest[0], latest[1]])),
             tap(([action, prevMacros, newMacros]: [Action, Macro[], Macro[]]) => {
                 const newMacro = findNewItem(prevMacros, newMacros);
                 const commands = ['/macro', newMacro.id];
